requirements.txt
streamlit
python-dotenv
openai
langchain
langchain-community
faiss-cpu
tiktoken